feat: Implement 'Apply to all' functionality in icon addition workflow
Enhanced the 'create-and-add-icon-to-folder' event
when users are managing multiple icon uploads
to support an 'Apply to all' option when adding batch icons to a folder.
This feature allow users to choose to skip or override existing icons
after the first prompt and repeat for all next icons.

The update includes:
- Conditional display of the 'Apply to all' checkbox only
when there is more than one icon being uploaded.
- Continuation or early return based on the user's choice and the 'Apply to all' selection.
- Introduction of an ACTIONS object for clearer representation of user choices.
- Change "showMessageBoxSynch" to "showMessageBox" to access built-in checkbox of message box
- Change "forEach" loop to "for ... of" in order to wait user input
without additioanl async function on forEach (already inside async function).